+++ src/enum.ts
@@ -1,0 +1,4 @@
+export enum Language {
+  chinese = 'zh',
+  english = 'en',
+}