+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # translate-Lpuena Change Log
 
+## v0.0.3 - 2023-08-05
+- New input box translation mode, shortcut key to start
+- Cancel automatic copy clipboard
+
 ## v0.0.2 - 2023-08-05
 - Add support for Chinese to English
 - Add support for clipboard  ✔